feat(canvas): clear canvas block content and remove mark on delete/backspace

+++ Sources/Editor/Commands/EditingCommands.swift
@@ -40,6 +40,9 @@ struct DeleteCharCommand: Command {
             return .succeeded
         }
         if editor.isCanvasModeActive {
+            if editor.clearCanvasBlockIfNeeded(saveSnapshot: false) {
+                return .succeeded
+            }
             editor.deleteCanvasCharacter()
             return .succeeded
         }
